+++ time-series/time-series-dsl/src/main/java/com/powsybl/timeseries/dsl/CalculatedTimeSeriesGroovyDslAstTransformation.java
@@ -6,43 +6,20 @@
  */
 package com.powsybl.timeseries.dsl;
 
-import org.codehaus.groovy.ast.ASTNode;
+import com.powsybl.dsl.AbstractPowsyblDslAstTransformation;
 import org.codehaus.groovy.ast.ClassCodeExpressionTransformer;
-import org.codehaus.groovy.ast.MethodNode;
-import org.codehaus.groovy.ast.ModuleNode;
 import org.codehaus.groovy.ast.expr.*;
-import org.codehaus.groovy.ast.stmt.BlockStatement;
 import org.codehaus.groovy.control.SourceUnit;
-import org.codehaus.groovy.transform.ASTTransformation;
 import org.codehaus.groovy.transform.GroovyASTTransformation;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.List;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
 @GroovyASTTransformation
-public class CalculatedTimeSeriesGroovyDslAstTransformation implements ASTTransformation {
+public class CalculatedTimeSeriesGroovyDslAstTransformation extends AbstractPowsyblDslAstTransformation {
 
-    protected static final Logger LOGGER = LoggerFactory.getLogger(CalculatedTimeSeriesGroovyDslAstTransformation.class);
-
-    protected void visit(SourceUnit sourceUnit, ClassCodeExpressionTransformer transformer) {
-        LOGGER.trace("Apply AST transformation");
-        ModuleNode ast = sourceUnit.getAST();
-        BlockStatement blockStatement = ast.getStatementBlock();
-
-        List<MethodNode> methods = ast.getMethods();
-        for (MethodNode methodNode : methods) {
-            methodNode.getCode().visit(transformer);
-        }
-
-        blockStatement.visit(transformer);
-    }
-
-    public void visit(ASTNode[] nodes, SourceUnit sourceUnit) {
-        visit(sourceUnit, new CustomClassCodeExpressionTransformer(sourceUnit));
+    public CalculatedTimeSeriesGroovyDslAstTransformation() {
+        super(CustomClassCodeExpressionTransformer::new);
     }
 
     static class CustomClassCodeExpressionTransformer extends ClassCodeExpressionTransformer {
